JAMES-2183 extract mailbox related stuff out of UserStepDef to MailboxStepDef
