+++ server/protocols/jmap-integration-testing/jmap-integration-testing-common/src/test/java/org/apache/james/jmap/methods/integration/cucumber/UserStepdefs.java
@@ -115,12 +115,6 @@ public class UserStepdefs {
         connectUser(username);
     }
     
-    @Given("^\"([^\"]*)\" has a mailbox \"([^\"]*)\"$")
-    public void createMailbox(String username, String mailbox) throws Throwable {
-        mainStepdefs.mailboxProbe.createMailbox(MailboxConstants.USER_NAMESPACE, username, mailbox);
-    }
-
-    
     @Given("^\"([^\"]*)\" is connected$")
     public void connectUser(String username) throws Throwable {
         AccessToken accessToken = getTokenForUser(username);
@@ -135,30 +129,6 @@ public class UserStepdefs {
 
             return HttpJmapAuthentication.authenticateJamesUser(mainStepdefs.baseUri(), user, password);
         });
-    }
-    
-    @Given("^\"([^\"]*)\" shares its mailbox \"([^\"]*)\" with \"([^\"]*)\"$")
-    public void shareMailbox(String owner, String mailbox, String shareTo) throws Throwable {
-        MailboxPath mailboxPath = MailboxPath.forUser(owner, mailbox);
-        MailboxACL.Rfc4314Rights rights = new MailboxACL.Rfc4314Rights(MailboxACL.Right.Lookup, MailboxACL.Right.Read);
-
-        mainStepdefs.aclProbe.addRights(mailboxPath, shareTo, rights);
-    }
-
-    @Given("^\"([^\"]*)\" shares its mailbox with write right \"([^\"]*)\" with \"([^\"]*)\"$")
-    public void shareMailboxWithWriteRight(String owner, String mailbox, String shareTo) throws Throwable {
-        MailboxPath mailboxPath = MailboxPath.forUser(owner, mailbox);
-        MailboxACL.Rfc4314Rights rights = new MailboxACL.Rfc4314Rights(MailboxACL.Right.Lookup, MailboxACL.Right.Read, MailboxACL.Right.Write);
-
-        mainStepdefs.aclProbe.addRights(mailboxPath, shareTo, rights);
-    }
-
-    @Given("^\"([^\"]*)\" does not share its mailbox \"([^\"]*)\" with \"([^\"]*)\"$")
-    public void notShareMailbox(String owner, String mailbox, String shareTo) throws Throwable {
-        MailboxPath mailboxPath = MailboxPath.forUser(owner, mailbox);
-        MailboxACL.Rfc4314Rights rights = new MailboxACL.Rfc4314Rights();
-
-        mainStepdefs.aclProbe.replaceRights(mailboxPath, shareTo, rights);
     }
 
     private String generatePassword(String username) {
